requirements.txt
Flask
qrcode[pil]
Pillow
python-dotenv